type mismatch

+++ lib/simd/sve/sve_acle.h
@@ -29,8 +29,8 @@
     //static inline svbool_t pg4(){return svptrue_pat_b16(SV_VL8);}
     static inline svbool_t pg4(){return svptrue_pat_b64(SV_VL2);}
     // exchange neighboring elements
-    static inline vec<double> tbl1(){
-        const vec<double> t = {1, 0, 3, 2, 5, 4, 7, 6};
+    static inline vec<uint64_t> tbl1(){
+        const vec<uint64_t> t = {1, 0, 3, 2, 5, 4, 7, 6};
         return t;
     }
     #elif (GEN_SIMD_WIDTH == 32u)
@@ -64,8 +64,8 @@
     //static inline svbool_t pg2(){return svptrue_pat_b16(SV_VL16);}
     static inline svbool_t pg2(){return svptrue_pat_b32(SV_VL8);}
     // exchange neighboring elements
-    static inline vec<float> tbl1(){
-        const vec<float> t = {1, 0, 3, 2, 5, 4, 7, 6, 9, 8, 11, 10, 13, 12, 15, 14};
+    static inline vec<uint32_t> tbl1(){
+        const vec<uint32_t> t = {1, 0, 3, 2, 5, 4, 7, 6, 9, 8, 11, 10, 13, 12, 15, 14};
         return t;
     }
     #elif (GEN_SIMD_WIDTH == 32u)
